feat: Add HomePage class
Description:
This commit adds the `HomePage` class, representing a page object for the home page of "theInternet" application. The class extends `AbstractPage` and provides methods to navigate to different pages by clicking on links. Three specific methods are implemented for navigating to the Form Authentication, Dropdown, and Key Presses pages.

Changes:
- Added `HomePage` class.
- Implemented methods to navigate to Form Authentication, Dropdown, and Key Presses pages.
- Utilized the `ExtendedWebElement` and `AbstractPage` classes for WebDriver interaction.